+++ packages/react/src/components/elements/input.tsx
@@ -1,33 +1,13 @@
 import React from "react";
 import { Input as ShadInput } from "../ui/input";
 import { cn } from "@/lib/utils";
-import { ElementAdornmentType } from "@lezzform/types/dist/shared";
+import { InputWithAdornmentProps } from "@/types";
 
-export type InputStyles<T = React.CSSProperties> = {
-  root: T;
-  input: T;
-  prefixAdornment?: Partial<Record<keyof ElementAdornmentType, T>>;
-  suffixAdornment?: InputStyles<T>["prefixAdornment"];
-};
-
-export type InputClassNames<T = string> = {
-  root: T;
-  input: T;
-  prefixAdornment?: Partial<Record<keyof ElementAdornmentType, T>>;
-  suffixAdornment?: InputClassNames<T>["prefixAdornment"];
-};
-
-interface Props extends React.InputHTMLAttributes<HTMLInputElement> {
+interface Props
+  extends React.InputHTMLAttributes<HTMLInputElement>,
+    InputWithAdornmentProps {
   isRequired?: boolean;
   label?: string;
-  styles?: Partial<InputStyles>;
-  classNames?: Partial<InputClassNames>;
-  prefixAdornment?: Partial<
-    Record<keyof ElementAdornmentType, React.JSX.Element>
-  >;
-  suffixAdornment?: Partial<
-    Record<keyof ElementAdornmentType, React.JSX.Element>
-  >;
 }
 
 export const Input = ({
@@ -37,18 +17,10 @@ export const Input = ({
   suffixAdornment,
   ...props
 }: Props) => {
-  const isUsingAdornment = Boolean(prefixAdornment) || Boolean(suffixAdornment);
-
-  if (!isUsingAdornment) {
-    return (
-      <ShadInput style={styles?.root} className={classNames?.root} {...props} />
-    );
-  }
-
   return (
     <div
       className={cn(
-        "lf-flex lf-items-center lf-h-10 lf-w-full lf-rounded-md lf-border lf-border-input lf-bg-background lf-text-sm file:lf-border file:lf-border-border file:lf-border-solid file:lf-rounded file:lf-outline-none file:lf-shadow-none file:lf-bg-muted file:lf-text-sm file:lf-font-medium file:lf-px-2 placeholder:lf-text-muted-foreground disabled:lf-cursor-not-allowed disabled:lf-opacity-50",
+        "lf-flex lf-items-center lf-h-10 lf-w-full lf-rounded-md lf-border lf-border-input lf-bg-background lf-text-sm file:lf-border file:lf-border-border file:lf-border-solid file:lf-rounded file:lf-outline-none file:lf-shadow-none file:lf-bg-muted file:lf-text-sm file:lf-font-medium file:lf-px-2 placeholder:lf-text-muted-foreground disabled:lf-cursor-not-allowed disabled:lf-opacity-50 focus-within:lf-outline-none  focus-within:lf-ring-2 focus-within:lf-ring-ring focus-within:lf-ring-offset-2",
         !prefixAdornment && "lf-pl-3",
         !suffixAdornment && "lf-pr-3",
         classNames?.root,