+++ packages/react/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./form";
 export * from "./generics";
+export * from "./element";